Add additional field functionality

+++ api/resources/additional_field.py
@@ -11,21 +11,33 @@ class AdditionalField(Resource):
 
         parser = reqparse.RequestParser()
         parser.add_argument("id", type=str, help="Additional field id is required", required=True)
-        parser.add_argument("field_name", type=str, help="Additional field id is required", required=True)
-        parser.add_argument("value", type=str, help="Additional field id is required", required=True)
-        parser.add_argument("record_id", type=str, help="Additional field id is required", required=True)
+        parser.add_argument("field_name", type=str, help="Additional field name is required", required=True)
+        parser.add_argument("value", type=str, help="Additional field value is required", required=True)
+        parser.add_argument("record_id", type=str, help="Additional field's record id is required", required=True)
+        parser.add_argument("nonce", type=str, help="Additional field's nonce is required", required=True)
+
 
         args = parser.parse_args()
+        record_with_recived_record_id = models.Record.query.get(args["record_id"])
+
+        if not record_with_recived_record_id:
+            return {"message": f"Record with recived record id doesn't exist"}, 404
+        elif current_user.id != record_with_recived_record_id.user_id:
+            return {"message": f"Record with id {record_with_recived_record_id.id} doesn't belong to the current user"}, 403
+        elif models.AdditionalField.query.get(args["id"]):
+            return {"message": f"Additional field with id '{args['id']}' already exist"}, 409
+        elif models.AdditionalField.query.filter_by(record_id=record_with_recived_record_id.id, field_name=args["field_name"]).all():
+            return {"message": f"Additional field with name '{args['field_name']}' already exist in this record"}, 409
+    
 
         additional_field = models.AdditionalField(id=args["id"], field_name=args["field_name"], value=args["value"], 
-                                record_id=args["record_id"])
+                                record_id=record_with_recived_record_id.id, nonce=args["nonce"])
 
-        record = models.Record.query.get(args["record_id"])
 
         db.session.add(additional_field)
         db.session.commit()
 
-        return {"message": f"Additional record field '{args['field_name']}' created successfully (record_name = '{record.name}',  user = '{current_user.email}')"}, 201
+        return {"message": f"Additional record field '{args['field_name']}' created successfully (record_name = '{record_with_recived_record_id.name}',  user = '{current_user.email}')"}, 201
     
     
     @token_required
@@ -38,12 +50,57 @@ class AdditionalField(Resource):
 
         args = parser.parse_args()
 
-        additional_field = models.AdditionalField.query.filter_by(id=args["id"]).all()
+        additional_field = models.AdditionalField.query.get(args["id"])
+       
+
+        if not additional_field:
+            return {"message": f"Additional field with id '{args['id']}' doesn't exist"}, 404
+
+        
+
         db.session.delete(additional_field)
         db.session.commit()
 
-        record = models.Record.query.get(additional_field.record_id)
-
-
-        return {"message": f"Additional record field '{additional_field.field_name}' deleted successfully (record_name = '{record.name}',  user = '{current_user.email}')"}, 200
+        return {"message": f"Additional field '{additional_field.field_name}' deleted successfully (user = '{current_user.email}')"}, 200
         
+
+    @token_required
+    def patch(self, current_user):
+        parser = reqparse.RequestParser()
+        parser.add_argument("id", type=str, help="Additional field id is required", required=True)
+        parser.add_argument("field_name", type=str)
+        parser.add_argument("value", type=str)
+        parser.add_argument("record_id", type=str, help="Additional field's record id is required", required=True)
+
+        args = parser.parse_args()
+
+        additional_field = models.AdditionalField.query.get(args["id"])
+        record_with_recived_record_id = models.Record.query.get(args["record_id"])
+
+
+        if not additional_field:
+            return {"message": f"Additional field with id '{args['id']}' doesn't exist "}, 404
+        elif not record_with_recived_record_id:
+            return {"message": f"Record with id '{record_with_recived_record_id.id}' doesn't exist "}, 404
+        elif current_user.id != record_with_recived_record_id.user_id:
+            return {"message": f"Record with id '{record_with_recived_record_id.id}' doesn't belong to the current user"}, 403
+        elif record_with_recived_record_id.id != additional_field.record_id:
+            return {"message": f"Additional field with id '{additional_field.id}' doesn't belong to the record with id '{record_with_recived_record_id.id}'"}, 409
+
+
+        additional_field_changed = False
+
+        if args["field_name"]:
+            additional_field.field_name = args["field_name"]
+            additional_field_changed = True
+
+        if args["value"]:
+            additional_field.value = args["value"]
+            additional_field_changed = True
+
+
+        if additional_field_changed:
+            db.session.commit()
+            return {"message": f"Changes for the additional field '{args['id']}' were successfully made"}, 200
+        else:
+            return {"message": f"Changes for the record '{args['id']}' weren't made because request body is empty"}, 400
